+++ babel.config.js
@@ -3,10 +3,7 @@ module.exports = {
         [
             '@babel/preset-env',
             {
-                targets: {
-                    node: 'current',
-                    edge: '18'
-                },
+                targets: "defaults",
             },
         ],
     ],