+++ src/route.ts
@@ -1,0 +1,11 @@
+class Route {
+	params = {};
+	path: string;
+	handlers: Array<Function>;
+	constructor(path: string, handlers: Array<Function>) {
+		this.path = path;
+		this.handlers = handlers;
+	}
+}
+
+export default Route;